refactor: rename TableService to MetaService + support both table,view
