+++ src/utils/index.ts
@@ -4,4 +4,3 @@ export * from './hooks';
 export * from './filter';
 export * from './row';
 export * from './sort';
-export * from './table';
